+++ example/lib/counter/counter_widget.dart
@@ -1,12 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:mobx_provider/src/mobx_providers/mobx_stateful_provider.dart';
+import 'package:mobx_provider/mobx_provider.dart';
+import 'package:mobx_provider/src/mobx_providers/mobx_widget_provider.dart';
+
 import 'counter_store.dart';
 
 class CounterWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MobxStatefulProvider<CounterStore>(
-      store: CounterStore(),
+      initFunction: (store) => store.increment(),
       builder: (context, store) {
         return Scaffold(
           floatingActionButton: FloatingActionButton(
@@ -19,7 +21,11 @@ class CounterWidget extends StatelessWidget {
               crossAxisAlignment: CrossAxisAlignment.center,
               mainAxisAlignment: MainAxisAlignment.center,
               children: <Widget>[
-                Text('counter value is ${store.counter}'),
+                MobxWidgetProvider<CounterStore>(
+                  builder: (context, store) {
+                    return Text('counter value is ${store.counter}');
+                  },
+                ),
               ],
             ),
           ),