+++ lib/example/counter_widget.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:mobx_provider/mobx_provider.dart';
 import 'package:mobx_provider/src/mobx_providers/mobx_stateful_provider.dart';
 import 'counter_store.dart';
 
@@ -6,7 +7,6 @@ class CounterWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MobxStatefulProvider<CounterStore>(
-      store: CounterStore(),
       builder: (context, store) {
         return Scaffold(
           floatingActionButton: FloatingActionButton(
@@ -18,7 +18,9 @@ class CounterWidget extends StatelessWidget {
           body: Center(
             child: Column(
               children: <Widget>[
-                Text('counter value is ${store.counter}'),
+                MobxWidgetProvider<CounterStore>(builder: (context, _) {
+                  return Text('counter value is ${store.counter}');
+                }),
               ],
             ),
           ),