+++ lib/mobx_provider.dart
@@ -3,4 +3,6 @@ library mobx_provider;
 export 'package:mobx_provider/mobx_provider.dart';
 export 'package:mobx_provider/src/mobx_base/mobx_base.dart';
 export 'package:mobx_provider/src/mobx_providers/mobx_stateful_provider.dart';
-export 'package:mobx_provider/src/mobx_providers/mobx_stateless_provider.dart';
+export 'package:mobx_provider/src/mobx_providers/mobx_stateful_observer.dart';
+export 'package:mobx_provider/src/mobx_providers/mobx_widget_provider.dart';
+export 'package:mobx_provider/src/mobx_providers/mobx_stateless_observer.dart';