+++ lib/src/mobx_providers/mobx_stateful_provider.dart
@@ -1,29 +1,21 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_mobx/flutter_mobx.dart';
-import 'package:mobx_provider/src/mobx_base/mobx_base.dart';
+import 'package:mobx_provider/mobx_provider.dart';
+import 'package:provider/provider.dart';
 
-///This is like the [Observer stateless Provider] but it contains [Stateful widget]
-///instead of [Stateless Widget], this allows you to call a function [initFunction]
-/// that you want to be invoked in the [init State]
-/// it also preserve the [state of the store] in hot reload
-/// if you press hot reload with the default [observer Provider] it'll lose its state
-/// when the [current widget] is removed from the widget tree the [StatefulObserverProvider] will
-/// call [disopse function] for the store
-/// this package is inspired from [Filled Stacks] package [view model provider]
-class MobxStatefulProvider<T extends MobxBase> extends StatefulObserverWidget {
-  final T store;
+///This Widget is integration between [Mobx] and [Provider] package
+///specify your [store] in the [provider] and this widget will call [provider.of<T>(context)]
+///to get it ,this widget doesn't observe the state of the store ,
+///this widget just exposes your store.
+///this widget is helpful when you want to get the store's functionality without updating the UI
+///like (counter.increment), use this widget with [MobxWidgetProvider] which observes the state of the store
+class MobxStatefulProvider<T extends MobxBase> extends StatefulWidget {
   final Widget Function(BuildContext context, T store) builder;
+
+  ///this function is called in the initState , can be used to fetch data from the api
   final void Function(T store) initFunction;
 
-  MobxStatefulProvider({
-    ///the name of your widget , that's used for debuging purposes only
-    ///it's the same as the name parameter in the [observer] widgets
-    String name,
-    Key key,
-    @required this.store,
-    @required this.builder,
-    this.initFunction,
-  }) : super(key: key, name: name);
+  const MobxStatefulProvider({Key key, this.builder, this.initFunction})
+      : super(key: key);
   @override
   _MobxStatefulProviderState<T> createState() =>
       _MobxStatefulProviderState<T>();
@@ -31,20 +23,20 @@ class MobxStatefulProvider<T extends MobxBase> extends StatefulObserverWidget {
 
 class _MobxStatefulProviderState<T extends MobxBase>
     extends State<MobxStatefulProvider<T>> {
-  T _store;
   @override
   void initState() {
+    if (widget.initFunction != null)
+      widget.initFunction(Provider.of<T>(context, listen: false));
     super.initState();
-    _store = widget.store;
-    if (widget.initFunction != null) widget.initFunction(_store);
   }
 
   @override
   void dispose() {
-    _store.dispose();
+    Provider.of<T>(context).dispose();
     super.dispose();
   }
 
   @override
-  Widget build(BuildContext context) => widget.builder(context, _store);
+  Widget build(BuildContext context) =>
+      widget.builder(context, Provider.of<T>(context));
 }