+++ CHANGELOG.md
@@ -7,3 +7,10 @@
 * add missing export statement
 * added example folder
 * improve reamde file
+
+## [0.0.2] 
+
+* change [MobxStatefulProvider] to [MobxStatefulObserver]
+* change [MobxStatelessProvider] to [MobxStatelessObserver]
+* add [MobxStatefulProvider] which provides the store to its children from the [Provider] package
+* add [MobxWidgetProvider] which observes the state of the store which is provided from [Provider] packge